Fix frequency undercompensation from truncation
Timestamp corrections are accumulated between batches of timestamps
processed by 'freqcd', and truncated to the nearest timestamp
resolution unit since floating point is not used. This results in a
gradual loss of correction accuracy across cumulative batches
(specifically undercompensation of timing) that results in a lower than
intended frequency compensation.

This is mitigated by using a 128-bit construct to store the overflow
correction instead, with the maximum 64 additional bits used for higher
resolution. This requires 128-bit integer support, which has been
available in gcc/clang for quite some time now on 64-bit platforms.
Performance impact is negligible due to batch processing of timestamps.

Fallback for lack of 128-bit support is the old legacy code, with a
suitable warning emitted during compile time. Future work to support
these platforms can be done as needed, by implementing a
128-bit-equivalent struct using two 64-bit signed/unsigned integers,
along with the add, mul, <<, >> operations.