docs: change versioning and commit styles
Searching for relevant commit messages for particular changes is
becoming tedious in recent days due to the lack of categorization of
commits. Migrating to SemVer with backward-compatible versioning
comparison entails upping the version to 2; the use of semantic
versioning will communicate more effectively whether the software is
safe to upgrade.