feat!: implement dynamic choice of FFT bins
This is useful under the case of large frequency offsets, where a larger
bin size is preferable to larger number of FFT bins (the latter
performance scales inversely to the frequency difference).

Since the exact frequency difference is usually not known, it is useful
to dynamically choose the FFT bins by virtue of changing the bin
resolution. This also allows us to effectively ignore the exact peak
significance calculation which behaves differently under low bin counts.