feat: allow fpfind to be interrupted with SIGINT
Because scipy has no hooks for interrupting the FFT calculations,
sending an interrupt signal often hangs while it waits for the internal
C routines to return.

This commit patches a SIGINT handler by running the FFT in a separate
process, at the expense of extra 20% runtime needed to shuttle data
between the process and the main thread.