fix!: ignore threshold check if no dt/df change detected
This fixes bugs that occur when no peak drift was detected at low
resolutions, but a drift at high resolutions. This is common, since at
low resolutions a drift in a peak near the origin may not be detectable,
i.e. dt = 0, df = 0, while dt != 0 and df != 0 at higher resolutions.